requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0
pillow>=9.0.0
pyautogui>=0.9.54
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
